added fn getListReviews which divided into modules
