+++ src/index.html
@@ -130,7 +130,6 @@
 <script src="js/form.js"></script>
 <script src="js/game.js"></script>
 <script src="js/main.js"></script>
-<script src="js/reviews.js"></script>
 
 </body>
 </html>
