+++ src/js/main.js
@@ -1,6 +1,7 @@
 'use strict';
 
 (function() {
+  var getListReviews = require('./list-reviews');
   var game = new window.Game(document.querySelector('.demo'));
   game.initializeLevelAndStart();
   game.setGameStatus(window.Game.Verdict.INTRO);
